Re-org tests
